% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimer.R
\docType{package}
\name{aimer}
\alias{aimer}
\alias{aimer-package}
\title{aimer: Amplified, Initially Marginal, Eigenvector Regression}
\description{
The aimer package implements the aimer algorithm as described in
\href{https://doi.org/10.1093/bioinformatics/btx265}{Ding and McDonald (2017)}.
The main goal is to use marginal regression to select a subset of coefficients, use
matrix approximation to estimate the principal components, and then extend those estimates
to the original predictor space before thresholding.
}
\details{
The package uses fast C++ routines to quickly perform matrix computations and cross validation for
selection of tuning parameters.

The package provides functions for estimating the model, choosing tuning parameters, and
generating simulated data as well as methods for prediction, plotting, and extraction.
}
\section{Estimation functions}{


\code{\link[=raimer]{raimer()}} Perform Amplified, Initially Marginal, Eigenvector Regression (AIMER) for fixed t, b, and d.
}

\section{Selecting tuning parameters}{

\code{\link[=findThresholdSelect]{findThresholdSelect()}} Find Optimal Threshold for Amplified, Initially Marginal,
Eigenvector Regression (AIMER) With Further Selection
}

\section{Simulation}{

}

\section{Methods}{

}

