% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v9_final.R
\name{approximateSuperPCA}
\alias{approximateSuperPCA}
\title{Approximate Supervised PCA for High-dimensional Regression}
\usage{
approximateSuperPCA(x, y, ncomp, nCov, nCov.select = NULL,
  PCAmethod = c("SPC", "SPC.lasso", "AIMER0", "AIMER"), tStats = NULL)
}
\arguments{
\item{x}{required, design matrix with dimension (n,p).}

\item{y}{required, response vector with dimension n.}

\item{ncomp}{required, number of components.}

\item{nCov}{required, number of covariates to compute SPC/SPC.lasso/AIMER0.}

\item{nCov.select}{required when PCAmethod="AIMER", can be < or = or > than nCov.}

\item{PCAmethod}{required, the method for approximate PCA.}

\item{tStats}{optional.}
}
\value{
a list with many components, includes much of the input, of class 'supervisedPCA'
\item{keep}{logical, which columns of \code{x} are used to compute SPC/AIMER0}
\item{tStats}{t-statistics via marginal regression for each covariate}
\item{gamhat}{regression coefficients on principal components}
\item{betahat}{regression coefficients on the kept \code{x} columns (note that order
has been changed)}
\item{ymean}{the mean of response \code{y}}
\item{xmeans}{the column means of design matrix \code{x}, for out-of-sample prediction}
\item{u}{approximated left singularvectors}
\item{d}{approximated sigularvalues}
\item{v}{approximated right sigularvectors}
}
\description{
performs SPC/SPC.lasso/AIMER0/AIMER for fixed \code{ncomp}, \code{nCov}
and \code{nCov.select}.
}
