% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{coef.aimer}
\alias{coef.aimer}
\title{returns the coefficients of an aimer model.}
\usage{
\method{coef}{aimer}(object, ...)
}
\arguments{
\item{object}{required, an aimer model.}

\item{...}{additional arguments, currently ignored.}
}
\value{
coefficient vector for the model
}
\description{
returns the coefficients of an aimer model.
}
