% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{coef.aimerCV}
\alias{coef.aimerCV}
\title{returns the coefficients of an aimerCV model.}
\usage{
\method{coef}{aimerCV}(object, ...)
}
\arguments{
\item{object}{required, an aimerCV model.}

\item{...}{additional arguments, currently ignored.}
}
\value{
coefficient vector for the model
}
\description{
returns the coefficients of an aimerCV model.
}
