% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{coefficients.aimer}
\alias{coefficients.aimer}
\title{returns the coefficients of an aimer model.}
\usage{
\method{coefficients}{aimer}(object, ...)
}
\arguments{
\item{object}{required, an aimer model.}
}
\value{
coefficient vector for the model
}
\description{
returns the coefficients of an aimer model.
}
