% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{coefficients.aimerCV}
\alias{coefficients.aimerCV}
\title{returns the coefficients of an aimerCV model.}
\usage{
\method{coefficients}{aimerCV}(object, ...)
}
\arguments{
\item{object}{required, an aimerCV model.}
}
\value{
coefficient vector for the model
}
\description{
returns the coefficients of an aimerCV model.
}
