% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v9_final.R
\name{estimateSuperPCA}
\alias{estimateSuperPCA}
\title{Estimate Supervised PCA for High-dimensional Regression}
\usage{
estimateSuperPCA(x, y, ncomps, nCovs = NULL,
  nCovs.min = ifelse(is.null(nCovs), max(ncomps) + 2, min(nCovs)),
  nCovs.max = ifelse(is.null(nCovs), nrow(x), max(nCovs)),
  nthresh = ifelse(is.null(nCovs), 25, length(nCovs)), nCovs.select = NULL,
  nCovs.min.select = ifelse(is.null(nCovs.select), max(ncomps) + 2,
  min(nCovs.select)), nCovs.max.select = ifelse(is.null(nCovs.select),
  nrow(x), max(nCovs.select)), nthresh.select = ifelse(is.null(nCovs.select),
  25, length(nCovs.select)), PCAmethod = c("SPC", "SPC.lasso", "AIMER0",
  "AIMER"), kfold = 10, progress = FALSE)
}
\arguments{
\item{x}{required, design matrix with dimension (n,p).}

\item{y}{required, response vector with dimension n.}

\item{ncomps}{required, denotes number of components, can be an integer
or a vector of integers.}

\item{nCovs}{optional, a vector of possible numbers of covariates.}

\item{nCovs.min}{optional, the smallest number of covariates, default as
max(\code{ncomps})+2.}

\item{nCovs.max}{optional, the largest number of covariates, default as
number of rows of \code{x}.}

\item{nthresh}{optional, how many \code{nCovs} to be tested, default as 25.}

\item{nCovs.select}{optional, a vector of possible numbers of covariates
in selection process}

\item{nCovs.min.select}{optional, the smallest number of covariates
in selection process, default as max(\code{ncomps})+2.}

\item{nCovs.max.select}{optional, the largest number of covariates
in selection process, default as number of rows of \code{x}.}

\item{nthresh.select}{optional, how many \code{nCovs.select} to be
tested in selection process, default as 25.}

\item{PCAmethod}{one of c("SPC", "SPC.lasso", "AIMER0", "AIMER"), defaults to SPC}

\item{kfold}{required, the number of k in kfold cross-validation,
default as 10.}

\item{progress}{logical, print how many folds have been calculated
and the the current number of covariates.}
}
\value{
two lists, the first is 'unfixTuning' of class 'supervisedPCACV'
which contains tuning parameter information, and the second is
'fixTuning' of class 'supervisedPCA' which is the estimated model.
}
\description{
the main function, automatically chooses best number of covariates,
best number of component for SPC/SPC.lasso/AIMER0, and best number of
covariates in selection step for AIMER by minimizing the CV
error, returns the estimated supervised PCA object with the
optimal values for tuning parameters.
}
