% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulTemp.R
\name{factorModelSim1}
\alias{factorModelSim1}
\title{Simulate some data from a sparse 2 factor model}
\usage{
factorModelSim1(n, p, p1 = 2, lambdas = c(5, 2), beta1 = 1, sig0 = 0.1,
  sig1 = 1)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of predictors}

\item{p1}{number of predictors which load onto the factors (requires 2*\code{p1} <= p)}

\item{lambdas}{relative weights of the factors}

\item{beta1}{coefficient of y on the first factor (the second is calculated automatically to make the marginal correlation between those predictors and the response 0)}

\item{sig0}{Standard deviation of the factor}

\item{sig1}{Standard deviation of the noise}
}
\value{
A list containing:
\enumerate{
\item \code{X}  --- n x p matrix of predictors
\item \code{Y}  --- response vector of length n
\item \code{theta}  --- regression coefficients on the predictors
\item \code{beta}  --- regression coefficients on the factors (first is \code{beta1})
\item \code{U}  --- p x 2 matrix of factors
\item \code{Lambda}  --- matrix of factor weights (as input)
\item \code{SigXY}  --- marginal correlation between columns of X and Y
}
}
\description{
Simulate some data from a sparse 2 factor model
}
\examples{
dat = factorModelSim1(100,1000,5,c(5,1),2,.1,.1)
}
