% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_v9_final.R
\name{findThreshold.select}
\alias{findThreshold.select}
\title{Find Optimal Threshold for Amplified, Initially Marginal,
Eigenvector Regression (AIMER) With Further Selection}
\usage{
findThreshold.select(x, y, ncomps, nCovs = NULL,
  nCovs.min = ifelse(is.null(nCovs), max(ncomps) + 2, min(nCovs)),
  nCovs.max = ifelse(is.null(nCovs), nrow(x), max(nCovs)),
  nthresh = ifelse(is.null(nCovs), 25, length(nCovs)), nCovs.select = NULL,
  nCovs.min.select = ifelse(is.null(nCovs.select), max(ncomps) + 2,
  min(nCovs.select)), nCovs.max.select = ifelse(is.null(nCovs.select),
  nrow(x), max(nCovs.select)), nthresh.select = ifelse(is.null(nCovs.select),
  25, length(nCovs.select)), kfold = 10, progress = FALSE)
}
\arguments{
\item{x}{required, design matrix with dimension (n,p).}

\item{y}{required, response vector with dimension n.}

\item{ncomps}{required, number of components, can be an integer or
a vector of integers.}

\item{nCovs}{optional, a vector of possible numbers of covariates.}

\item{nCovs.min}{optional, the smallest number of covariates, default
as max(\code{ncomps})+2.}

\item{nCovs.max}{optional, the largest number of covariates, default
as number of rows of \code{x}.}

\item{nthresh}{optional, how many \code{nCovs} to be tested, default as 25.}

\item{nCovs.select}{optional, a vector of possible numbers of covariates
in selection process.}

\item{nCovs.min.select}{optional, the smallest number of covariates
in selection process, default as max(\code{ncomps})+2.}

\item{nCovs.max.select}{optional, the largest number of covariates
in selection process, default as number of rows of \code{x}.}

\item{nthresh.select}{optional, how many \code{nCovs.select} to
be tested in selection process, default as 25.}

\item{kfold}{required, the number of k in kfold cross-validation,
default as 10.}

\item{progress}{logical, print how many nCov/folds have been calculated.}
}
\value{
an object of class 'supervisedPCACV', a list with the following components
\item{nCov.select.best}{the best number of covariates in selection process
which gives smallest mse in cross-validation}
\item{nCov.best}{the best number of covariates which gives smallest mse
in cross-validation}
\item{ncomp.best}{the best number of components which gives smallest mse
in cross-validation}
\item{nCovs.select}{all the tested numbers of covariates in selection process}
\item{ncomps}{all the tested ncomps}
\item{nCovs}{all the tested numbers of covariates}
\item{CVmse}{mse in cross-validation with respect to each value
of \code{nCovs}, \code{ncomps}, \code{kfold}}
\item{mse}{average mse in cross-validation over k folds}
}
\description{
find the optimal number of covariates, number of components, and number of covariates
in selection process for AIMER method using kfold cross-validation.
}
