% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raimer.R
\name{findThresholdAIMER0}
\alias{findThresholdAIMER0}
\title{Find Optimal Threshold for Amplified, Initially Marginal,
Eigenvector Regression (AIMER) Without Further Selection}
\usage{
findThresholdAIMER0(X, y, nComps, nCovs = NULL,
  nCovs.min = ifelse(is.null(nCovs), max(nComps) + 2, min(nCovs)),
  nCovs.max = ifelse(is.null(nCovs), nrow(X), max(nCovs)),
  nthresh = ifelse(is.null(nCovs), 25, length(nCovs)), kfold = 10)
}
\arguments{
\item{y}{required, response vector with dimension n.}

\item{nCovs}{optional, a vector of possible numbers of covariates.}

\item{nCovs.min}{optional, the smallest number of covariates, default
as max(\code{ncomps})+2.}

\item{nCovs.max}{optional, the largest number of covariates, default
as number of rows of \code{x}.}

\item{nthresh}{optional, how many \code{nCovs} to be tested, default as 25.}

\item{kfold}{required, the number of k in kfold cross-validation,
default as 10.}

\item{x}{required, design matrix with dimension (n,p).}

\item{ncomps}{required, number of components, can be an integer or
a vector of integers.}
}
\value{
an object of class 'supervisedPCACV', a list with the following components
\item{nCov.best}{the best number of covariates which gives smallest mse
in cross-validation}
\item{ncomp.best}{the best number of components which gives smallest mse
in cross-validation}
\item{ncomps}{all the tested ncomps}
\item{nCovs}{all the tested numbers of covariates}
\item{CVmse}{mse in cross-validation with respect to each value
of \code{nCovs}, \code{ncomps}, \code{kfold}}
\item{mse}{average mse in cross-validation over k folds}
}
\description{
find the optimal number of covariates and number of components using
kfold cross-validation for AIMER0
}
