% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{fitted.aimer}
\alias{fitted.aimer}
\title{returns the fitted values of an aimer model.}
\usage{
\method{fitted}{aimer}(object, ...)
}
\arguments{
\item{object}{required, an aimer model.}
}
\value{
fitted values of the original data (vector of length n).
}
\description{
returns the fitted values of an aimer model.
}
