% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{fitted.aimerCV}
\alias{fitted.aimerCV}
\title{returns the fitted values of an aimerCV model.}
\usage{
\method{fitted}{aimerCV}(object, ...)
}
\arguments{
\item{object}{required, an aimerCV model.}
}
\value{
fitted values of the original data (vector of length n).
}
\description{
returns the fitted values of an aimerCV model.
}
