% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{predict.aimer}
\alias{predict.aimer}
\title{predicts new values for an aimer model.}
\usage{
\method{predict}{aimer}(object, newdata, ...)
}
\arguments{
\item{object}{required, a model generated by the aimer function.}

\item{newdata}{required, a new data matrix to predict with.}

\item{...}{additional arguments, currently ignored.}
}
\value{
predicted value vector of length nrow(newdata).
}
\description{
predicts new values for an aimer model.
}
