% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{predict.aimerCV}
\alias{predict.aimerCV}
\title{predicts new values for an aimerCV model.}
\usage{
\method{predict}{aimerCV}(object, newdata, ...)
}
\arguments{
\item{object}{required, a model generated by the aimerCV function.}

\item{newdata}{required, a new data matrix to predict with.}

\item{...}{additional arguments, currently ignored.}
}
\value{
predicted value vector of length nrow(newdata).
}
\description{
predicts new values for an aimerCV model.
}
