% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predictSuperPCA}
\alias{predictSuperPCA}
\title{Predict Supervised PCA}
\usage{
predictSuperPCA(obj, newx = NULL, PCAmethod = c("SPC", "SPC.lasso",
  "AIMER0", "AIMER"))
}
\arguments{
\item{obj}{an object of class 'supervisedPCA' as output by approximateSuperPCA.}

\item{newx}{an optional matrix of points for predictions, if missing, then uses
original design matrix.}

\item{PCAmethod}{required.}
}
\value{
predictions for newx design matrix
}
\description{
use this function to predict response for feature data.
}
