% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raimer.R
\name{raimer}
\alias{raimer}
\title{performs Amplified, Initially Marginal, Eigenvector Regression (AIMER) for fixed t, b, and d}
\usage{
raimer(X, y, nCovs, nCovsSelect, nComps)
}
\arguments{
\item{X}{required, design matrix with dimension (n, p).}

\item{y}{required, response vector with dimension n.}

\item{nCovs}{required, number of highest marginally correlated covariates to use.}

\item{nCovsSelect}{required, desired number of non-zero coefficients.}

\item{nComps}{required, number of principal components.}
}
\value{
coefficient vector of length p.
}
\description{
performs Amplified, Initially Marginal, Eigenvector Regression (AIMER) for fixed t, b, and d
}
