% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{residuals.aimer}
\alias{residuals.aimer}
\title{returns the residuals of an aimer model.}
\usage{
\method{residuals}{aimer}(object, ...)
}
\arguments{
\item{object}{required, an aimer model.}
}
\value{
residuals of the original data (vector of length n).
}
\description{
returns the residuals of an aimer model.
}
