% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{residuals.aimerCV}
\alias{residuals.aimerCV}
\title{returns the residuals of an aimerCV model.}
\usage{
\method{residuals}{aimerCV}(object, ...)
}
\arguments{
\item{object}{required, an aimerCV model.}
}
\value{
residuals of the original data (vector of length n).
}
\description{
returns the residuals of an aimerCV model.
}
