% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIMER.R
\docType{package}
\name{aimer}
\alias{aimer}
\alias{aimer-package}
\title{aimer: Approxima}
\description{
The aimer package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

