% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{plot.aimer}
\alias{plot.aimer}
\title{plots the residuals in terms of the fitted values for an aimer model}
\usage{
\method{plot}{aimer}(object, ...)
}
\arguments{
\item{object}{required, an aimer model.}
}
\value{
void.
}
\description{
plots the residuals in terms of the fitted values for an aimer model
}
