% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aimerMethods.R
\name{plot.aimerCV}
\alias{plot.aimerCV}
\title{creates a heatmap of each value of ncomps tested in an aimerCV model}
\usage{
\method{plot}{aimerCV}(x, ...)
}
\arguments{
\item{x}{required, an aimerCV model.}

\item{...}{additional arguments, currently ignored}
}
\value{
void.
}
\description{
creates a heatmap of each value of ncomps tested in an aimerCV model
}
