% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raimer.R
\name{raimer}
\alias{raimer}
\title{performs Amplified, Initially Marginal, Eigenvector Regression (AIMER) for fixed t, b, and d}
\usage{
raimer(X, y, t, b, d)
}
\arguments{
\item{X}{required, design matrix with dimension (n, p).}

\item{y}{required, response vector with dimension n.}

\item{t}{required, threshold for marginal coefficients.}

\item{b}{required, threshold for output coefficients.}

\item{d}{required, number of principal components.}
}
\value{
coefficient vector of length p.
}
\description{
performs Amplified, Initially Marginal, Eigenvector Regression (AIMER) for fixed t, b, and d
}
